Demote a couple of dependencies from SNAPSHOT to release versions
This is all in line with the goal to have reproducible builds already.

Technically, it does not achieve that because the referenced release
versions all have parent POMs at SNAPSHOT versions. In practice, however,
the referenced projects should not put the maintenance burden on the
Fiji maintainer, pretending that "community" means "all the work is
done by the maintainer". Therefore, practically speaking, there will
not be any reproducible build of those projects anyway, at least not
until the project owners change their maintenance model.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>