Bump the last two non-TrakEM2 components to release versions
Now the TrakEM2-related things only need to be moved to their own parent
POM and we're ready for the next beta release version of pom-fiji.

Note: We did not reach our goal of fully reproducible builds yet: some
of the releases we cut *still* depend on -SNAPSHOT versions. This is
something we will need to fix one by one, and the maven-enforcer-plugin
(which is configured by the pom-scijava version we depend on
transitively) will help with that.  Signed-off-by: Johannes Schindelin
<johannes.schindelin@gmx.de>