Remove invalid release versions from the POM
Unfortunately, the first non-beta pom-fiji release references invalid
components: its Bill of Materials lists some components that are
nominally release versions, but invalid ones because the components
themselves depend on SNAPSHOT versions, making the end result
unpredictable and hence unreliable.

Let's fix that bug in the pom-fiji and remove those components until a
stage when they are available in correct release versions. It will be
very easy for the maintainer to include the components again by
reverting this here commit selectively.

The components in question have been identified by this shell script
(due to BSD-sed quirks – which requires a newline after a curly brace –
it runs only on Linux, but it was written for this one-off tasks only,
anyway):

	#!/bin/sh

	for av in $(sed -n '/Fiji.sub.projects/,/\/properties/s/.*<\([^>\/]*\)\.version>\([^<]*\).*/\1:\2/p' < pom.xml);
	do
		a=${av%:*}
		v=${av#*:}
		g="$(sed -n '/<groupId>/{N;s/.*<groupId>\([^<]*\).*<artifactId>'"$a"'.*/\1/p}' < pom.xml)"
		url=http://maven.imagej.net/service/local/repo_groups/public/content
		url=$url/$(echo "$g" | tr . /)
		url=$url/$a/$v/$a-$v.pom

		printf 'Testing %s...\r' "$a"
		mkdir -p .git/tmp
		snapshots="$(cd .git/tmp &&
		 curl -s $url > pom.xml &&
		 mvn dependency:tree 2>&1 |
		 sed -n '/^.INFO. --- maven-dependency-plugin/,/^.INFO -*$/s/.* \([^ ]*-SNAPSHOT\).*/\1/p')"

		test -z "$snapshots" ||
		printf 'Depends on SNAPSHOT versions: %s\n\n%s\n' "$g:$a:$v" "$snapshots"
	done

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>