Transfer BDV-based projects to core fiji org
These projects are not part of core BDV, but rather merely depend on
it for their operation. They were once in the fiji org, but moved to
bigdataviewer because they needed to utilize the pom-bigdataviewer BOM.

However, the decision was made for pom-trakem2 to stop managing the
versions of its downstream components (see
https://github.com/trakem2/TrakEM2/commit/afc201b6798332cfa4fbcf3147cf38aae6a9d0d6),
with their version management returning to pom-fiji (see
https://github.com/fiji/pom-fiji/commit/5eeb909bb0586e9816658abe4eb17abd8ac0aa66).
Hence, for consistency, these BDV-downstream components are also being
returned to the fiji org and pom-fiji for version management.

They will import pom-bigdataviewer's BOM using scope import, and
extend pom-fiji, like the rest of the standard Fiji components.

This also avoids future potential issues relating to BOM inheritance
from multiple Fiji sub-projects at the same time -- i.e., it would
now be easy for either of these BDV-based components to also depend
on components of TrakEM2.